+++ pulse-binding-simple/CHANGELOG.md
@@ -1,8 +1,9 @@
 # [unreleased]
 
+ * Changed the return type of `Simple::get_latency()` to pass error codes to the caller on error.
  * Tweaked the implementation of `Simple::get_latency()` such that in the hypothetical situation (it
    is not clear whether this ever happens) whereby the underlying C function indicates success yet
-   returns `PA_USEC_INVALID`, this will now result in `None` being returned, rather than
+   returns `PA_USEC_INVALID`, this will now result in an error being returned, rather than
    `MicroSeconds(PA_USEC_INVALID)`.
 
 # 2.24.1 (August 29th, 2021)